requirements.txt
anyio==4.0.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzdata==2023.3
tzlocal==5.1
